Move month/year and month aggregation into helpers
